+++ packages/runtime/src/lib/constants.ts
@@ -40,6 +40,10 @@ export const MAX_INNER_TRANSACTIONS = 16;
 export const ALGORAND_MAX_LOGS_COUNT = 32;
 export const ALGORAND_MAX_LOGS_LENGTH = 1024;
 
+export const publicKeyLength = 32;
+export const proofLength = 80;
+export const seedLength = 32;
+
 export const MAX_ALGORAND_ACCOUNT_ASSETS = 1000;
 export const MAX_ALGORAND_ACCOUNT_CREATED_APPS = 10;
 
@@ -55,6 +59,9 @@ export const MAX_LOCAL_SCHEMA_ENTRIES = 16;
 // https://github.com/algorand/go-algorand/blob/bd5a00092c8a63dba8314b97851e46ff247cf7c1/data/transactions/logic/eval.go#L1302
 export const MAX_INPUT_BYTE_LEN = 64;
 export const MAX_OUTPUT_BYTE_LEN = 128;
+
+export const MaxTxnLife = 1000;
+export const BlockFinalisationTime = 4n; // block finalisation time in seconds truncated down
 
 export const ZERO_ADDRESS = new Uint8Array(32);
 export const ZERO_ADDRESS_STR = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAY5HFKQ";
@@ -477,6 +484,7 @@ OpGasCost[7] = {
 	ed25519verify_bare: 1900,
 	ecdsa_verify: 2500,
 	ecdsa_pk_decompress: 2400,
+	vrf_verify : 5700,
 };
 
 export const enum MathOp {
@@ -517,6 +525,16 @@ export const json_refTypes = {
 	JSONString: "JSONString",
 	JSONUint64: "JSONUint64",
 	JSONObject: "JSONObject",
+};
+
+export enum blockFieldTypes {
+	BlkTimestamp = "BlkTimestamp",
+	BlkSeed = "BlkSeed",
+};
+
+export enum vrfVerifyFieldTypes {
+    VrfAlgorand = "VrfAlgorand",
+	VrfStandard = "VrfStandard",
 };
 
 export enum TxFieldEnum {